Atualizar completamente pronto

+++ arquivos/src/usuarios/usuarios.js
@@ -1,27 +1,27 @@
-var teste = false
-window.addEventListener('DOMContentLoaded', () => {
-    const nomeUsuario = sessionStorage.getItem('guardarNomes');
-    if (nomeUsuario) {
-      console.log(`Usuário ${nomeUsuario} já está logado.`);
-      teste = true;
-      console.log("tem coisa guardada")
-    }
-    if (teste === true) {
-      console.log("tqa entrando")
-      fetch(`http://localhost:3000/usuarios/${nomeUsuario}`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      })
-        .then((response) => response.json())
-        .then((data) => {
-          let guardarNome = data.nomeUsuario;
-          sessionStorage.setItem("guardarNome", JSON.stringify(guardarNome));
+var teste = false;
+window.addEventListener("DOMContentLoaded", () => {
+  const nomeUsuario = sessionStorage.getItem("guardarNomes");
+  if (nomeUsuario) {
+    console.log(`Usuário ${nomeUsuario} já está logado.`);
+    teste = true;
+    console.log("tem coisa guardada");
+  }
+  if (teste === true) {
+    console.log("tqa entrando");
+    fetch(`http://localhost:3000/usuarios/nomeUsuario/${nomeUsuario}`, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+      },
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        let guardarNome = data.nomeUsuario;
+        sessionStorage.setItem("guardarNome", JSON.stringify(guardarNome));
 
-          console.log("bem vindo");
+        console.log("bem vindo");
 
-          let idUsuarioD = data.id;
+        let idUsuarioD = data.id;
 
         //   const sidebar = document.createElement("div");
         //   sidebar.id = "sidebar";
@@ -41,12 +41,12 @@ window.addEventListener('DOMContentLoaded', () => {
         // `;
         //   document.body.appendChild(sidebar);
 
-          toggleSidebar();
-          closeDialog();
+        toggleSidebar();
+        closeDialog();
 
-          carregarListaUsuarios();
-        });
-    }
+        carregarListaUsuarios();
+      });
+  }
 });
 
 async function initMap() {
@@ -66,7 +66,6 @@ async function initMap() {
   const styleControl = document.getElementById("controleSelecao");
   map.controls.push(styleControl);
   const styleSelector = document.getElementById("theme");
-
 
   styleSelector.addEventListener("click", () => {
     if (styleSelector.checked) {
@@ -121,7 +120,7 @@ async function initMap() {
                 <h5 class="cardTitulo">${posto.nomePosto}</h5>
                 <h6 class="cardSubtitulo mb-2 text-muted">Endereço</h6>
                 <p class="cardTexto">Cidade: ${posto.enderecoPosto}</p>
-                <p class="cardTexto">Cidade: ${posto.ruaPosto}</p>
+                <p class="cardTexto">Rua: ${posto.ruaPosto}</p>
               </div>
             </div>
             <div class="card">
@@ -423,8 +422,8 @@ document.addEventListener("DOMContentLoaded", function () {
     });
 });
 
-function modoCor() { }
-function modoCor() { }
+function modoCor() {}
+function modoCor() {}
 
 function openDialog() {
   let customDialog = document.getElementById("loginForm");
@@ -570,16 +569,16 @@ function adicionarUsuario() {
 
 function toggleSidebar() {
   const navBar = document.querySelector("nav"),
-        menuBtns = document.querySelectorAll(".menu-icon"),
-        overlay = document.querySelector(".overlaySidebar");
-      menuBtns.forEach((menuBtn) => {
-        menuBtn.addEventListener("click", () => {
-          navBar.classList.toggle("open");
-        });
-      });
-      overlay.addEventListener("click", () => {
-        navBar.classList.remove("open");
-      });
+    menuBtns = document.querySelectorAll(".menu-icon"),
+    overlay = document.querySelector(".overlaySidebar");
+  menuBtns.forEach((menuBtn) => {
+    menuBtn.addEventListener("click", () => {
+      navBar.classList.toggle("open");
+    });
+  });
+  overlay.addEventListener("click", () => {
+    navBar.classList.remove("open");
+  });
 }
 
 function toggleSideLar() {
@@ -593,15 +592,15 @@ function toggleSideLar() {
     overlay.parentNode.removeChild(overlay);
   }
 
-  sessionStorage.removeItem('guardarNomes');
-  sessionStorage.removeItem('guardarNome');
+  sessionStorage.removeItem("guardarNomes");
+  sessionStorage.removeItem("guardarNome");
 
   var listaItem = document.querySelector(".usuario");
   let userList = document.getElementById("formPerfil");
 
   userList.removeChild(listaItem);
 
-  var sidebar = document.querySelector("nav")
+  var sidebar = document.querySelector("nav");
 
   document.body.removeChild(sidebar);
 }
@@ -624,7 +623,7 @@ loginForm.addEventListener("submit", async (e) => {
     .map((byte) => byte.toString(16).padStart(2, "0"))
     .join("");
 
-  fetch(`http://localhost:3000/usuarios/${userName}`, {
+  fetch(`http://localhost:3000/usuarios/nomeUsuario/${userName}`, {
     method: "GET",
     headers: {
       "Content-Type": "application/json",
@@ -638,25 +637,31 @@ loginForm.addEventListener("submit", async (e) => {
 
         sessionStorage.setItem("guardarNomes", userName);
 
+        let guardarSenha = data.senha;
+
+        sessionStorage.setItem("guardarSenha", JSON.stringify(guardarSenha));
+
+        sessionStorage.setItem("guardarSenhas", hashSenha);
+
         console.log("bem vindo");
 
         idUsuarioD = data.id;
         sessionStorage.setItem("idUsuarioDs", idUsuarioD);
 
-      //   const sidebar = document.createElement("nav");
-      //   sidebar.id = "sidebar";
-      //   sidebar.classList.add("sidebar");
-      //   sidebar.innerHTML = `
-      //   <div>
-      //     <ul>
-      //       <li><div class="intern-div-list"><i class="bi bi-person"><a onclick="dialogPerfil();" class="botaoPerfil" id="botaoPerfil"">Perfil</div></i></li>
-      //       <li><div class="intern-div-list"><i class="bi bi-fuel-pump-fill"><a href="postoteste.html">Postos</a></div></i></li>
-      //       <li id="listaUsuarios"><div class="intern-div-list"><i class="bi bi-people"><a href="usuarios.html">Usuários</a></div></i></li>
-      //       <li><div class="intern-div-list"><a href="#">Item 4</a></div></li>
-      //       <li><div class="intern-div-list"><a href="#">Item 5</a></div></li>
-      //     </ul>
-      //   </div>
-      // `;
+        //   const sidebar = document.createElement("nav");
+        //   sidebar.id = "sidebar";
+        //   sidebar.classList.add("sidebar");
+        //   sidebar.innerHTML = `
+        //   <div>
+        //     <ul>
+        //       <li><div class="intern-div-list"><i class="bi bi-person"><a onclick="dialogPerfil();" class="botaoPerfil" id="botaoPerfil"">Perfil</div></i></li>
+        //       <li><div class="intern-div-list"><i class="bi bi-fuel-pump-fill"><a href="postoteste.html">Postos</a></div></i></li>
+        //       <li id="listaUsuarios"><div class="intern-div-list"><i class="bi bi-people"><a href="usuarios.html">Usuários</a></div></i></li>
+        //       <li><div class="intern-div-list"><a href="#">Item 4</a></div></li>
+        //       <li><div class="intern-div-list"><a href="#">Item 5</a></div></li>
+        //     </ul>
+        //   </div>
+        // `;
         // document.body.appendChild(sidebar);
 
         toggleSidebar();
@@ -669,10 +674,9 @@ loginForm.addEventListener("submit", async (e) => {
     });
 });
 
-
 function carregarListaUsuarios() {
-  const idUsuarioD = sessionStorage.getItem('idUsuarioDs');
-  console.log(idUsuarioD)
+  const idUsuarioD = sessionStorage.getItem("idUsuarioDs");
+  console.log(idUsuarioD);
 
   fetch(`http://localhost:3000/usuarios/id/${idUsuarioD}`)
     .then((response) => response.json())
@@ -799,28 +803,70 @@ function fecharDialogNome() {
   dialog.style.display = "none";
 }
 
-function atualizarNomeUsuario() {
+async function atualizarNomeUsuario() {
+  const antigoNomeUsuario = document.getElementById("antigoNomeUsuario").value;
   const novoNomeUsuario = document.getElementById("novoNomeUsuario").value;
+  const confirmeSenhaNome = document.getElementById("confirmeSenhaNome").value;
   const idUsuario = document.getElementById("userId").value;
 
-  let guardarNome = novoNomeUsuario;
-  sessionStorage.setItem("guardarNome", JSON.stringify(guardarNome));
+  const encoder = new TextEncoder();
+  const data = encoder.encode(confirmeSenhaNome);
+  const hashBuffer = await crypto.subtle.digest("SHA-256", data);
+  const hashArray = Array.from(new Uint8Array(hashBuffer));
+  const hashSenha = hashArray
+    .map((byte) => byte.toString(16).padStart(2, "0"))
+    .join("");
 
-  fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
-    method: "PUT",
+  fetch(`http://localhost:3000/usuarios/nomeUsuario/${antigoNomeUsuario}`, {
+    method: "GET",
     headers: {
       "Content-Type": "application/json",
     },
-    body: JSON.stringify({
-      nomeUsuario: novoNomeUsuario,
-    }),
   })
     .then((response) => response.json())
     .then((data) => {
-      console.log("Usuário atualizado com sucesso:", data);
-      carregarListaUsuarios();
-    })
-    .catch((error) => console.error("Erro ao atualizar o usuário:", error));
+      if (
+        antigoNomeUsuario === "" ||
+        novoNomeUsuario === "" ||
+        confirmeSenhaNome === ""
+      ) {
+        alert("Preencha todos os campos");
+      } else if (novoNomeUsuario === "admin" || novoNomeUsuario === "Admin") {
+        alert("Nome de usuário não pode ser admin");
+      } else if (
+        !verificarNome(antigoNomeUsuario) ||
+        !verificarNome(novoNomeUsuario)
+      ) {
+        alert("Nome de usuário inválido");
+      } else if (data.nomeUsuario !== antigoNomeUsuario) {
+        console.log(data.nomeUsuario);
+
+        alert("Nome de usuário não encontrado");
+      } else if (data.senha !== hashSenha) {
+        alert("Senha não confere");
+      } else {
+        let guardarNome = novoNomeUsuario;
+        sessionStorage.setItem("guardarNome", JSON.stringify(guardarNome));
+
+        fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            nomeUsuario: novoNomeUsuario,
+          }),
+        })
+          .then((response) => response.json())
+          .then((data) => {
+            console.log("Usuário atualizado com sucesso:", data);
+            carregarListaUsuarios();
+          })
+          .catch((error) =>
+            console.error("Erro ao atualizar o usuário:", error)
+          );
+      }
+    });
 }
 
 // Atualizar email
@@ -837,25 +883,72 @@ function fecharDialogEmail() {
   dialog.style.display = "none";
 }
 
-function atualizarEmailUsuario() {
+async function atualizarEmailUsuario() {
+  const antigoEmailUsuario = document.getElementById("emailUsuario").value;
   const novoEmailUsuario = document.getElementById("novoEmailUsuario").value;
+  const confirmeSenhaEmail =
+    document.getElementById("confirmeSenhaEmail").value;
   const idUsuario = document.getElementById("userId").value;
 
-  fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
-    method: "PUT",
+  const encoder = new TextEncoder();
+  const data = encoder.encode(confirmeSenhaEmail);
+  const hashBuffer = await crypto.subtle.digest("SHA-256", data);
+  const hashArray = Array.from(new Uint8Array(hashBuffer));
+  const hashSenha = hashArray
+    .map((byte) => byte.toString(16).padStart(2, "0"))
+    .join("");
+
+  fetch(`http://localhost:3000/usuarios/email/${antigoEmailUsuario}`, {
+    method: "GET",
     headers: {
       "Content-Type": "application/json",
     },
-    body: JSON.stringify({
-      email: novoEmailUsuario,
-    }),
   })
     .then((response) => response.json())
     .then((data) => {
-      console.log("Usuário atualizado com sucesso:", data);
-      carregarListaUsuarios();
-    })
-    .catch((error) => console.error("Erro ao atualizar o usuário:", error));
+      if (
+        antigoEmailUsuario === "" ||
+        novoEmailUsuario === "" ||
+        confirmeSenhaEmail === ""
+      ) {
+        console.log(antigoEmailUsuario);
+        console.log(novoEmailUsuario);
+        console.log(confirmeSenhaEmail);
+        console.log(idUsuario);
+
+        alert("Preencha todos os campos corretamente");
+      } else if (
+        !verificarEmail(novoEmailUsuario) ||
+        !verificarEmail(antigoEmailUsuario)
+      ) {
+        alert("Email inválido");
+      } else if (data.email !== antigoEmailUsuario) {
+        alert("Email não existe");
+      } else if (data.senha !== hashSenha) {
+        alert("Senha incorreta");
+      } else {
+        let guardarEmail = novoEmailUsuario;
+        sessionStorage.setItem("guardarEmail", JSON.stringify(guardarEmail));
+
+        fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            email: novoEmailUsuario,
+          }),
+        })
+          .then((response) => response.json())
+          .then((data) => {
+            console.log("Usuário atualizado com sucesso:", data);
+            carregarListaUsuarios();
+          })
+          .catch((error) =>
+            console.error("Erro ao atualizar o usuário:", error)
+          );
+      }
+    });
 }
 
 // Atualizar celular
@@ -872,27 +965,69 @@ function fecharDialogTelefone() {
   dialog.style.display = "none";
 }
 
-function atualizarCelularUsuario() {
-  const novoTelefoneUsuario = document.getElementById(
-    "novoTelefoneUsuario"
-  ).value;
+async function atualizarCelularUsuario() {
+  const antigoTelefoneUsuario =document.getElementById("telefoneUsuario").value;
+  const novoTelefoneUsuario = document.getElementById("novoTelefoneUsuario").value;
+  const confirmeSenhaTelefone = document.getElementById("confirmeSenhaTelefone").value;
   const idUsuario = document.getElementById("userId").value;
 
-  fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
-    method: "PUT",
+  const encoder = new TextEncoder();
+  const data = encoder.encode(confirmeSenhaTelefone);
+  const hashBuffer = await crypto.subtle.digest("SHA-256", data);
+  const hashArray = Array.from(new Uint8Array(hashBuffer));
+  const hashSenha = hashArray
+    .map((byte) => byte.toString(16).padStart(2, "0"))
+    .join("");
+
+  fetch(`http://localhost:3000/usuarios/telefone/${antigoTelefoneUsuario}`, {
+    method: "GET",
     headers: {
       "Content-Type": "application/json",
     },
-    body: JSON.stringify({
-      telefone: novoTelefoneUsuario,
-    }),
   })
     .then((response) => response.json())
     .then((data) => {
-      console.log("Usuário atualizado com sucesso:", data);
-      carregarListaUsuarios();
-    })
-    .catch((error) => console.error("Erro ao atualizar o usuário:", error));
+      if (
+        antigoTelefoneUsuario === "" ||
+        novoTelefoneUsuario === "" ||
+        confirmeSenhaTelefone === ""
+      ) {
+        alert("Preencha todos os campos corretamente");
+      } else if (
+        !verificarTelefone(antigoTelefoneUsuario) ||
+        !verificarTelefone(novoTelefoneUsuario)
+      ) {
+        alert("Telefone inválido");
+      } else if (data.telefone !== antigoTelefoneUsuario) {
+        alert("Telefone atual não corresponde ao telefone do usuário");
+      } else if (data.senha !== hashSenha) {
+        alert("Senha incorreta");
+      } else {
+        let guardarTelefone = novoTelefoneUsuario;
+        sessionStorage.setItem(
+          "guardarTelefone",
+          JSON.stringify(guardarTelefone)
+        );
+
+        fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          body: JSON.stringify({
+            telefone: novoTelefoneUsuario,
+          }),
+        })
+          .then((response) => response.json())
+          .then((data) => {
+            console.log("Usuário atualizado com sucesso:", data);
+            carregarListaUsuarios();
+          })
+          .catch((error) =>
+            console.error("Erro ao atualizar o usuário:", error)
+          );
+      }
+    });
 }
 
 // Atualizar senha
@@ -909,25 +1044,58 @@ function fecharDialogSenha() {
   dialog.style.display = "none";
 }
 
-function atualizarSenhaUsuario() {
+async function atualizarSenhaUsuario() {
+  const antigaSenhaUsuario = document.getElementById("senhaUsuario").value;
   const novaSenhaUsuario = document.getElementById("novaSenhaUsuario").value;
+  const confirmeSenha = document.getElementById("confirmeSenha").value;
   const idUsuario = document.getElementById("userId").value;
 
-  fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
-    method: "PUT",
+  const encoder = new TextEncoder();
+  const data = encoder.encode(antigaSenhaUsuario);
+  const hashBuffer = await crypto.subtle.digest("SHA-256", data);
+  const hashArray = Array.from(new Uint8Array(hashBuffer));
+  const hashSenha = hashArray
+    .map((byte) => byte.toString(16).padStart(2, "0"))
+    .join("");
+
+  fetch(`http://localhost:3000/usuarios/senha/${hashSenha}`, {
+    method: "GET",
     headers: {
       "Content-Type": "application/json",
     },
-    body: JSON.stringify({
-      senha: novaSenhaUsuario,
-    }),
   })
     .then((response) => response.json())
-    .then((data) => {
-      console.log("Usuário atualizado com sucesso:", data);
-      carregarListaUsuarios();
+    .then(async (data) => {
+      if(antigaSenhaUsuario === "" || novaSenhaUsuario === "" || confirmeSenha === ""){
+        alert("Preencha todos os campos");
+      } else if (data.senha !== hashSenha){
+        alert("Senha antiga incorreta!");
+      } else if (novaSenhaUsuario !== confirmeSenha){
+        alert("As senhas não conferem!");
+      } else {
+
+        let guardarsenha = confirmeSenha;
+        sessionStorage.setItem("guardarsenha",JSON.stringify(guardarsenha));
+
+          fetch(`http://localhost:3000/usuarios/${idUsuario}`, {
+            method: "PUT",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({
+              senha: confirmeSenha,
+            }),
+          })
+            .then((response) => response.json())
+            .then((data) => {
+              console.log("Usuário atualizado com sucesso:", data);
+              carregarListaUsuarios();
+            })
+            .catch((error) => console.error("Erro ao atualizar o usuário:", error));
+      }
+    
     })
-    .catch((error) => console.error("Erro ao atualizar o usuário:", error));
+
 }
 
 // function closePerfilDialog() {
